chore: move ActionModelWithNullifiers into models
