+++ web/src/stores/actionStore.ts
@@ -1,11 +1,5 @@
-import { ActionModel, NullifierModel } from "src/lib/models";
+import { ActionModelWithNullifiers } from "src/lib/models";
 import { create } from "zustand";
-
-export interface ActionModelWithNullifiers extends ActionModel {
-  nullifiers: Array<
-    Pick<NullifierModel, "id" | "nullifier_hash" | "created_at">
-  >;
-}
 
 export type IActionStore = {
   actions: Array<ActionModelWithNullifiers>;
